// src/models/summonerLeague.ts
export interface SummonerLeague {
    leagueData:LeagueData[]
}
export interface LeagueData{
    leagueId: string;
    summonerId: string;
    summonerName: string;
    queueType: string;
    tier: string;
    rank: string;
    leaguePoints: number;
    wins: number;
    losses: number;
    hotStreak: boolean;
    veteran: boolean;
    freshBlood: boolean;
    inactive: boolean;
}